Add files via upload
 App Type Distribution (Pie Chart)

The pie chart illustrates the proportion of apps that are Free versus Paid on the Play Store. It is clearly visible that a vast majority (around 93%) of apps are free, with only a small portion (around 7%) being paid. This highlights the common business model on the Play Store, where most apps are free to download and rely on ads, in-app purchases, or freemium upgrades for monetization.