Add files via upload
Installs by Category (Bar Chart)

This bar chart shows the total number of installs per app category. Social, Game, and Communication categories dominate in terms of total installs. This insight is important for app developers and businesses because it highlights which categories have the largest user base and potential for high engagement and monetization.